Refactoring interactive endpoint, moving the detail of registration approval out to util file and bringing in the processing of the Find dialog.
